Add multierr.Is(err, target)
This is useful when the caller receives `error` type, and would like to
know if the error is of particular interest, in case it is a multierror.

Imagine a program `usergroups` that emits many lines to stdout, and is
used like this:

```
set -o pipefail
echo "First 20 users:"
usergroups users | head -20
echo "First 20 groups:"
usergroups groups | head -20
```

After 20 lines have been printed, stdout will be closed, and subsequent
`os.Stdout.Write(...)` calls will fail. If we are checking for errors
from the `Write` calls, the application should terminate with non-zero
exit code. That way, the script will terminate after 3 lines and will
ignore the poor groups.

However, in this case, `| head` is a user's choice, thus a closed
stdout/stderr is usually expected, and should not fail the program.

What to do? Close the program normally if stdout/stderr cannot be
written to. To detect it, it's often convenient to use
`multierr.Append`. Callee:

```
type ErrBadSink = errors.New("bad sink")
func execute(w io.Writer) error {
    if _, err := io.WriteString(w, "---header---"); err != nil {
        return multierr.Append(err, ErrBadSink)
    }
    if err := <...>; err != nil {
        return fmt.Errorf("other kind of error: %w")
    }
}
```

Caller would like to know whether the error is `ErrBadSink`. In my case,
if it is, it should terminate the program differently:

```
func main() {
    if err := execute(os.Stdout); err != nil {
        if multierr.Is(err, ErrBadSink) {
            // writing to stdout failed. Probably caused by a closed
            // file descriptor, say, because of `| head`. Terminate
            // successfully.
            os.Exit(0)
        }
        fmt.Fprintf(os.Stderr, "failed to run command: %s", err)
    }
}
```

This change allows testing the plain `error` types against
`multierr.Is`.